Extracted DraggableTicket.
